fix error for script run on busybox64.exe

+++ tools/check_added_ut.sh
@@ -21,6 +21,7 @@ if [ -z ${BRANCH} ]; then
 fi
 
 export CI_SKIP_CPP_TEST=OFF
+SYSTEM=`uname -s`
 if [ "$SYSTEM" == "Linux" ];then
     PADDLE_ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}")/../" && pwd )"
 elif ["$SYSTEM" == "Windows_NT" ];then
@@ -28,13 +29,14 @@ elif ["$SYSTEM" == "Windows_NT" ];then
 fi
 CURDIR=`pwd`
 cd $PADDLE_ROOT
-cp $PADDLE_ROOT/paddle/scripts/paddle_build.sh $PADDLE_ROOT/paddle/scripts/paddle_build_pre.sh
+if [ "$SYSTEM" == "Linux" ];then
+    cp $PADDLE_ROOT/paddle/scripts/paddle_build.sh $PADDLE_ROOT/paddle/scripts/paddle_build_pre.sh
+fi
 CURBRANCH=`git rev-parse --abbrev-ref HEAD`
 echo $CURBRANCH
 git checkout -b prec_added_ut upstream/${BRANCH}
 mkdir prec_build
 cd prec_build
-SYSTEM=`uname -s`
 if [ "$SYSTEM" == "Linux" ];then
     bash $PADDLE_ROOT/paddle/scripts/paddle_build_pre.sh cmake_gen_in_current_dir >prebuild.log 2>&1
 elif ["$SYSTEM" == "Windows_NT" ];then
@@ -50,7 +52,9 @@ sort pr-ut |uniq -d > $PADDLE_ROOT/duplicate_ut
 echo "New-UT:"
 cat $PADDLE_ROOT/added_ut
 rm -rf prec_build
-rm $PADDLE_ROOT/br-ut $PADDLE_ROOT/pr-ut $PADDLE_ROOT/paddle/scripts/paddle_build_pre.sh
+if [ "$SYSTEM" == "Linux" ];then
+    rm $PADDLE_ROOT/br-ut $PADDLE_ROOT/pr-ut $PADDLE_ROOT/paddle/scripts/paddle_build_pre.sh
+fi
 git checkout $CURBRANCH
 echo $CURBRANCH
 git branch -D prec_added_ut
